Remove float arithmetic from chat/embedding cost computation
chat.py::_record_usage converted model_catalog.input_per_million /
output_per_million (NUMERIC columns, Decimal via asyncpg, meant to be
whole tomans per the Wallet Mapped[int] type hint) to float before
computing cost — binary floating-point imprecision in a money path.
Use int() instead: an exact, lossless conversion that keeps the
existing half-up rounding formula unchanged.

embeddings.py::_bill_embedding_usage used true division (tokens * 100
/ 1_000_000) then truncated, both introducing a float and rounding
down instead of half-up. Switched to integer floor division with the
same (+ 500_000) // 1_000_000 half-up convention already used in
services/metering.py::_part.

+++ backend/chat.py
@@ -482,8 +482,8 @@ async def _record_usage(session: AsyncSession, uid: int, payload: dict[str, Any]
         logger.warning(f"_record_usage price lookup failed model={model} uid={uid}: {e}")
 
     if price_row:
-        inp_rate = float(price_row.input_per_million or 0)
-        out_rate = float(price_row.output_per_million or 0)
+        inp_rate = int(price_row.input_per_million or 0)
+        out_rate = int(price_row.output_per_million or 0)
         cost = max(1, int((input_tokens * inp_rate + output_tokens * out_rate + 500_000) // 1_000_000))
     else:
         cost = max(1, total_tokens // 1000)
